[Refactor] 파일 경로 변경
