+++ src/main/java/com/backend/moamoa/domain/user/dto/request/TokenRequest.java
@@ -1,4 +1,4 @@
-package com.backend.moamoa.domain.user.oauth.dto.request;
+package com.backend.moamoa.domain.user.dto.request;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
@@ -15,11 +15,11 @@ import javax.validation.constraints.NotNull;
 public class TokenRequest {
 
     @NotNull
-    @ApiModelProperty(value = "만료된 accessToken", required = true)
+    @ApiModelProperty(value = "만료된 accessToken", example = "", required = true)
     private String accessToken;
 
     @NotNull
-    @ApiModelProperty(value = "유저의 refreshToken", required = true)
+    @ApiModelProperty(value = "유저의 refreshToken", example = "", required = true)
     private String refreshToken;
 
 }
