+++ src/main/java/com/backend/moamoa/domain/user/entity/enums/Gender.java
@@ -1,0 +1,5 @@
+package com.backend.moamoa.domain.user.entity.enums;
+
+public enum Gender {
+    MAN, WOMAN
+}
