--- src/main/java/com/backend/moamoa/domain/user/enums/Gender.java
@@ -1,5 +1,0 @@
-package com.backend.moamoa.domain.user.enums;
-
-public enum Gender {
-    MAN, WOMAN
-}
