+++ src/main/java/com/backend/moamoa/global/bean/security/JwtConfig.java
@@ -1,18 +1,20 @@
-package com.backend.moamoa.global.config;
+package com.backend.moamoa.global.bean.security;
 
-import com.backend.moamoa.domain.user.oauth.provider.AuthTokenProvider;
+import com.backend.moamoa.domain.user.oauth.token.JwtProvider;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
+@RequiredArgsConstructor
 public class JwtConfig {
+
     @Value("${jwt.secret}")
     private String secret;
 
     @Bean
-    public AuthTokenProvider jwtProvider() {
-        return new AuthTokenProvider(secret);
+    public JwtProvider provider() {
+        return new JwtProvider(secret);
     }
-
 }
